+++ src/previewdef/focustree/loader.ts
@@ -1,8 +1,9 @@
-import { ContentLoader, LoadResultOD, Dependency, LoaderSession } from "../../util/loader";
+import { ContentLoader, LoadResultOD, Dependency, LoaderSession, LoadResult, mergeInLoadResult } from "../../util/loader";
 import { FocusTree, getFocusTree } from "./schema";
 import { parseHoi4File } from "../../hoiformat/hoiparser";
 import { localize } from "../../util/i18n";
-import { uniq } from "lodash";
+import { uniq, flatten, chain } from "lodash";
+import { error as debugError } from "../../util/debug";
 
 export interface FocusTreeLoaderResult {
     focusTrees: FocusTree[];
@@ -16,16 +17,36 @@ export class FocusTreeLoader extends ContentLoader<FocusTreeLoaderResult> {
         if (error || (content === undefined)) {
             throw error;
         }
+        
+        const focusTreeDependencies = dependencies.filter(d => d.type === 'focus').map(d => d.path);
+        const focusTreeDepFiles = (await Promise.all(focusTreeDependencies.map(async (dep) => {
+            try {
+                const focusTreeDepLoader = this.loaderDependencies.getOrCreate(dep, k => session.createOrGetCachedLoader(k, FocusTreeLoader), FocusTreeLoader);
+                return await focusTreeDepLoader.load(session);
+            } catch (e) {
+                debugError(e);
+                return undefined;
+            }
+        }))).filter((v): v is LoadResult<FocusTreeLoaderResult> => !!v);
 
-        const gfxDependencies = dependencies.filter(d => d.type === 'gfx').map(d => d.path);
-        const focusTrees = getFocusTree(parseHoi4File(content, localize('infile', 'In file {0}:\n', this.file)));
+        const sharedFocusTrees = chain(focusTreeDepFiles)
+            .flatMap(f => f.result.focusTrees)
+            .filter(ft => ft.isSharedFocues)
+            .value();
+
+        const focusTrees = getFocusTree(parseHoi4File(content, localize('infile', 'In file {0}:\n', this.file)), sharedFocusTrees, this.file);
+
+        const gfxDependencies = [
+            ...dependencies.filter(d => d.type === 'gfx').map(d => d.path),
+            ...flatten(focusTreeDepFiles.map(f => f.result.gfxFiles))
+        ];
 
         return {
             result: {
                 focusTrees,
                 gfxFiles: uniq([...gfxDependencies, focusesGFX]),
             },
-            dependencies: uniq([this.file, focusesGFX, ...gfxDependencies]),
+            dependencies: uniq([this.file, focusesGFX, ...gfxDependencies, ...focusTreeDependencies]),
         };
     }